# Conflicts:
#	myDB/db.lck
#	myDB/log/log.ctrl
#	myDB/log/log1.dat
#	myDB/log/logmirror.ctrl
#	myDB/seg0/c121.dat
#	myDB/seg0/c230.dat
#	myDB/seg0/c241.dat
#	myDB/seg0/c81.dat
#	src/main/java/edu/wpi/teamname/services/algo/PathFinder.java
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/dbex.lck
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/log/log1.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c10.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c130.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c141.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c180.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c1a1.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c1b1.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c1f1.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c20.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c31.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c380.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c391.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c3a1.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c3b1.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c400.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c41.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c411.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c421.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c51.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c81.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/c90.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/ca1.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/seg0/cf0.dat
#	src/main/java/edu/wpi/teamname/services/database/DerbyDB/service.properties
#	src/main/java/edu/wpi/teamname/views/MapController.java
#	src/main/resources/edu/wpi/teamname/views/map.fxml

+++ src/main/java/edu/wpi/teamname/services/algo/Node.java
@@ -6,18 +6,18 @@ import lombok.NonNull;
 
 public abstract class Node<T extends Node<T>> implements Comparable<Node<T>> {
   public class Link {
-    int _distance;
+    double _distance;
     Node<T> _other;
 
-    public Link(Node<T> other, int distance) {
+    public Link(Node<T> other, double distance) {
       this._distance = distance;
       this._other = other;
     }
   }
 
   T value;
-  int _localGoal;
-  int _globalGoal;
+  double _localGoal;
+  double _globalGoal;
   boolean _seen;
   Node<T> _parent;
   List<Link> _neighbors;
@@ -35,11 +35,11 @@ public abstract class Node<T extends Node<T>> implements Comparable<Node<T>> {
     this._parent = null;
   }
 
-  public void addNeighbor(Node<T> other, int distance) {
+  public void addNeighbor(Node<T> other, double distance) {
     this._neighbors.add(new Link(other, distance));
   }
 
-  abstract int heuristic(Node<T> other);
+  abstract double heuristic(Node<T> other);
 
   @Override
   public int compareTo(@NonNull Node<T> Other) {
